# Conflicts:
#	HMCL/src/main/java/org/jackhuang/hmcl/ui/download/LocalModpackPage.java
#	HMCL/src/main/java/org/jackhuang/hmcl/ui/download/RemoteModpackPage.java
#	HMCL/src/main/resources/assets/fxml/download/modpack.fxml

+++ HMCL/src/main/java/org/jackhuang/hmcl/setting/Theme.java
@@ -23,17 +23,21 @@ import com.google.gson.stream.JsonWriter;
 import javafx.beans.binding.Bindings;
 import javafx.beans.binding.ObjectBinding;
 import javafx.scene.paint.Color;
+import org.jackhuang.hmcl.util.Lang;
 import org.jackhuang.hmcl.util.Logging;
-import org.jackhuang.hmcl.util.ResourceNotFoundError;
 import org.jackhuang.hmcl.util.io.FileUtils;
 import org.jackhuang.hmcl.util.io.IOUtils;
-import org.jackhuang.hmcl.util.javafx.BindingMapping;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
+import java.util.Locale;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.logging.Level;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import static org.jackhuang.hmcl.setting.ConfigHolder.config;
 
@@ -49,6 +53,39 @@ public class Theme {
             Color.web("#9C27B0"), // purple
             Color.web("#B71C1C")  // red
     };
+
+    private static Charset cssCharset;
+
+    private static Charset getCssCharset() {
+        if (cssCharset != null)
+            return cssCharset;
+
+        Charset defaultCharset = Charset.defaultCharset();
+        if (defaultCharset != StandardCharsets.UTF_8) {
+            // https://bugs.openjdk.org/browse/JDK-8279328
+            // For JavaFX 17 or earlier, native encoding should be used
+            String jfxVersion = System.getProperty("javafx.version");
+            if (jfxVersion != null) {
+                Matcher matcher = Pattern.compile("^(?<version>[0-9]+)").matcher(jfxVersion);
+                if (matcher.find()) {
+                    int v = Lang.parseInt(matcher.group(), -1);
+                    if (v >= 18) {
+                        cssCharset = StandardCharsets.UTF_8;
+                    }
+                }
+            }
+        }
+
+        if (cssCharset == null)
+            cssCharset = defaultCharset;
+
+        return cssCharset;
+    }
+
+    public static Theme getTheme() {
+        Theme theme = config().getTheme();
+        return theme == null ? BLUE : theme;
+    }
 
     private final Color paint;
     private final String color;
@@ -73,7 +110,7 @@ public class Theme {
     }
 
     public boolean isLight() {
-        return Color.web(color).grayscale().getRed() >= 0.5;
+        return paint.grayscale().getRed() >= 0.5;
     }
 
     public Color getForegroundColor() {
@@ -81,33 +118,32 @@ public class Theme {
     }
 
     public String[] getStylesheets(String overrideFontFamily) {
-        Color textFill = getForegroundColor();
+        String css = "/assets/css/blue.css";
 
         String fontFamily = System.getProperty("hmcl.font.override", overrideFontFamily);
 
-        String css;
-        try {
-            File temp = File.createTempFile("hmcl", ".css");
-            FileUtils.writeText(temp, IOUtils.readFullyAsString(ResourceNotFoundError.getResourceAsStream("/assets/css/custom.css"))
-                    .replace("%base-color%", color)
-                    .replace("%base-red%", Integer.toString((int)Math.ceil(paint.getRed() * 256)))
-                    .replace("%base-green%", Integer.toString((int)Math.ceil(paint.getGreen() * 256)))
-                    .replace("%base-blue%", Integer.toString((int)Math.ceil(paint.getBlue() * 256)))
-                    .replace("%base-rippler-color%", String.format("rgba(%d, %d, %d, 0.3)", (int) Math.ceil(paint.getRed() * 256), (int) Math.ceil(paint.getGreen() * 256), (int) Math.ceil(paint.getBlue() * 256)))
-                    .replace("%disabled-font-color%", String.format("rgba(%d, %d, %d, 0.7)", (int) Math.ceil(textFill.getRed() * 256), (int) Math.ceil(textFill.getGreen() * 256), (int) Math.ceil(textFill.getBlue() * 256)))
-                    .replace("%font-color%", getColorDisplayName(getForegroundColor()))
-                    .replace("%font%", Optional.ofNullable(fontFamily).map(f -> "-fx-font-family: \"" + f + "\";").orElse("")));
-            temp.deleteOnExit();
-            css = temp.toURI().toString();
-        } catch (IOException | NullPointerException e) {
-            Logging.LOG.log(Level.SEVERE, "Unable to create theme stylesheet. Fallback to blue theme.", e);
-            css = "/assets/css/blue.css";
+        if (fontFamily != null || !this.color.equalsIgnoreCase(BLUE.color)) {
+            Color textFill = getForegroundColor();
+            try {
+                File temp = File.createTempFile("hmcl", ".css");
+                String themeText = IOUtils.readFullyAsString(Theme.class.getResourceAsStream("/assets/css/custom.css"))
+                        .replace("%base-color%", color)
+                        .replace("%base-red%", Integer.toString((int) Math.ceil(paint.getRed() * 256)))
+                        .replace("%base-green%", Integer.toString((int) Math.ceil(paint.getGreen() * 256)))
+                        .replace("%base-blue%", Integer.toString((int) Math.ceil(paint.getBlue() * 256)))
+                        .replace("%base-rippler-color%", String.format("rgba(%d, %d, %d, 0.3)", (int) Math.ceil(paint.getRed() * 256), (int) Math.ceil(paint.getGreen() * 256), (int) Math.ceil(paint.getBlue() * 256)))
+                        .replace("%disabled-font-color%", String.format("rgba(%d, %d, %d, 0.7)", (int) Math.ceil(textFill.getRed() * 256), (int) Math.ceil(textFill.getGreen() * 256), (int) Math.ceil(textFill.getBlue() * 256)))
+                        .replace("%font-color%", getColorDisplayName(getForegroundColor()))
+                        .replace("%font%", Optional.ofNullable(fontFamily).map(f -> "-fx-font-family: \"" + f + "\";").orElse(""));
+                FileUtils.writeText(temp, themeText, getCssCharset());
+                temp.deleteOnExit();
+                css = temp.toURI().toString();
+            } catch (IOException | NullPointerException e) {
+                Logging.LOG.log(Level.SEVERE, "Unable to create theme stylesheet. Fallback to blue theme.", e);
+            }
         }
 
-        return new String[]{
-                css,
-                "/assets/css/root.css"
-        };
+        return new String[]{css, "/assets/css/root.css"};
     }
 
     public static Theme custom(String color) {
@@ -119,50 +155,69 @@ public class Theme {
     public static Optional<Theme> getTheme(String name) {
         if (name == null)
             return Optional.empty();
-        else if (name.equalsIgnoreCase("blue"))
-            return Optional.of(custom("#5C6BC0"));
-        else if (name.equalsIgnoreCase("darker_blue"))
-            return Optional.of(custom("#283593"));
-        else if (name.equalsIgnoreCase("green"))
-            return Optional.of(custom("#43A047"));
-        else if (name.equalsIgnoreCase("orange"))
-            return Optional.of(custom("#E67E22"));
-        else if (name.equalsIgnoreCase("purple"))
-            return Optional.of(custom("#9C27B0"));
-        else if (name.equalsIgnoreCase("red"))
-            return Optional.of(custom("#F44336"));
-
-        if (name.startsWith("#"))
+        else if (name.startsWith("#"))
             try {
                 Color.web(name);
                 return Optional.of(custom(name));
             } catch (IllegalArgumentException ignore) {
             }
+        else {
+            String color = null;
+            switch (name.toLowerCase(Locale.ROOT)) {
+                case "blue":
+                    return Optional.of(BLUE);
+                case "darker_blue":
+                    color = "#283593";
+                    break;
+                case "green":
+                    color = "#43A047";
+                    break;
+                case "orange":
+                    color = "#E67E22";
+                    break;
+                case "purple":
+                    color = "#9C27B0";
+                    break;
+                case "red":
+                    color = "#F44336";
+            }
+            if (color != null)
+                return Optional.of(new Theme(name, color));
+        }
 
         return Optional.empty();
     }
 
     public static String getColorDisplayName(Color c) {
-        return c != null ? String.format("#%02x%02x%02x", Math.round(c.getRed() * 255.0D), Math.round(c.getGreen() * 255.0D), Math.round(c.getBlue() * 255.0D)).toUpperCase() : null;
+        return c != null ? String.format("#%02x%02x%02x", Math.round(c.getRed() * 255.0D), Math.round(c.getGreen() * 255.0D), Math.round(c.getBlue() * 255.0D)).toUpperCase(Locale.ROOT) : null;
     }
 
+    private static final ObjectBinding<Color> BLACK_FILL = Bindings.createObjectBinding(() -> BLACK);
+    private static final ObjectBinding<Color> WHITE_FILL = Bindings.createObjectBinding(() -> Color.WHITE);
+    private static ObjectBinding<Color> FOREGROUND_FILL;
+
     public static ObjectBinding<Color> foregroundFillBinding() {
-        return BindingMapping.of(config().themeProperty())
-                .map(Theme::getForegroundColor);
+        if (FOREGROUND_FILL == null)
+            FOREGROUND_FILL = Bindings.createObjectBinding(
+                    () -> Theme.getTheme().getForegroundColor(),
+                    config().themeProperty()
+            );
+
+        return FOREGROUND_FILL;
     }
 
     public static ObjectBinding<Color> blackFillBinding() {
-        return Bindings.createObjectBinding(() -> BLACK);
+        return BLACK_FILL;
     }
 
     public static ObjectBinding<Color> whiteFillBinding() {
-        return Bindings.createObjectBinding(() -> Color.WHITE);
+        return WHITE_FILL;
     }
 
     public static class TypeAdapter extends com.google.gson.TypeAdapter<Theme> {
         @Override
         public void write(JsonWriter out, Theme value) throws IOException {
-            out.value(value.getName().toLowerCase());
+            out.value(value.getName().toLowerCase(Locale.ROOT));
         }
 
         @Override
